Moved sanitizeInteraction into its own file
